refactor: move mock property to mock directory
