+++ src/components/lessorIncome-page/lessorDashboard.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { PropertyDetail } from '@/components/lesseeHome-page/PropertyDetail';
-import { mockData } from '@/components/example/mockdata';
+import { mockData } from '@/mocks/mockProperty';
 import { PropertyI } from '@/types/property';
 
 export default function LessorDashboard() {
